moved logic from HangmanFrame.java to GamePanel.java
