+++ src/main/java/com/mycompany/hangman/controller/WordController.java
@@ -5,7 +5,7 @@
  */
 package com.mycompany.hangman.controller;
 
-import com.mycompany.hangman.gui.HangmanFrame;
+import com.mycompany.hangman.gui.GamePanel;
 import com.mycompany.hangman.model.HangmanGame;
 import com.mycompany.hangman.model.Resetable;
 import java.awt.event.ActionEvent;
@@ -18,10 +18,10 @@ import java.awt.event.ActionListener;
 public class WordController extends AbstractController implements ActionListener, Resetable
 {
 
-    private final HangmanFrame view;
+    private final GamePanel  view;
     private final HangmanGame model;
 
-    public WordController(HangmanFrame view, HangmanGame model)
+    public WordController(GamePanel view, HangmanGame model)
     {
         this.view = view;
         this.model = model;
