+++ src/main/java/com/mycompany/hangman/gui/HangmanFrame.java
@@ -4,21 +4,13 @@
  */
 package com.mycompany.hangman.gui;
 
-import com.mycompany.hangman.drawing.DrawPanel;
-import com.mycompany.hangman.model.Character;
-import com.mycompany.hangman.model.HangmanGame;
-import java.awt.event.ActionListener;
-import java.beans.PropertyChangeEvent;
-import java.util.List;
 import javax.swing.*;
-import javax.swing.text.AbstractDocument;
-import javax.swing.text.DefaultCaret;
 
 /**
  *
  * @author Cory
  */
-public class HangmanFrame extends JFrame implements View
+public class HangmanFrame extends JFrame
 {
     /**
      * Creates new form NewJFrame
@@ -26,13 +18,8 @@ public class HangmanFrame extends JFrame implements View
     public HangmanFrame()
     {
         initComponents();
-        this.outputConsole.setEditable(false);
-        this.jScrollPane1.setWheelScrollingEnabled(true);
-        DefaultCaret caret = (DefaultCaret)this.outputConsole.getCaret();
-        caret.setUpdatePolicy(DefaultCaret.ALWAYS_UPDATE);
         setVisible(true);
         setTitle("Hangman");
-        ((AbstractDocument)this.guess.getDocument()).setDocumentFilter(new ValidGuessDocFilter());
     }
 
     /**
@@ -45,52 +32,9 @@ public class HangmanFrame extends JFrame implements View
     private void initComponents()
     {
 
-        guess = new javax.swing.JTextField();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        outputConsole = new javax.swing.JTextArea();
-        jButton1 = new javax.swing.JButton();
-        guessedLettersLabel = new javax.swing.JLabel();
-        guessedLetters = new javax.swing.JLabel();
-        drawPanel = new com.mycompany.hangman.drawing.DrawPanel();
-        wordToGuess = new javax.swing.JLabel();
+        gamePanel = new com.mycompany.hangman.gui.GamePanel();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        guess.setText("Enter guess here ");
-        guess.addFocusListener(new java.awt.event.FocusAdapter()
-        {
-            public void focusGained(java.awt.event.FocusEvent evt)
-            {
-                guessFocusGained(evt);
-            }
-        });
-
-        jScrollPane1.setAutoscrolls(true);
-
-        outputConsole.setColumns(20);
-        outputConsole.setRows(5);
-        jScrollPane1.setViewportView(outputConsole);
-
-        jButton1.setLabel("OK");
-
-        guessedLettersLabel.setText("Guessed Letters:");
-
-        guessedLetters.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
-        guessedLetters.setText("jLabel2");
-
-        javax.swing.GroupLayout drawPanelLayout = new javax.swing.GroupLayout(drawPanel);
-        drawPanel.setLayout(drawPanelLayout);
-        drawPanelLayout.setHorizontalGroup(
-            drawPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 198, Short.MAX_VALUE)
-        );
-        drawPanelLayout.setVerticalGroup(
-            drawPanelLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 198, Short.MAX_VALUE)
-        );
-
-        wordToGuess.setFont(new java.awt.Font("Tahoma", 0, 24)); // NOI18N
-        wordToGuess.setText("jLabel1");
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -98,51 +42,19 @@ public class HangmanFrame extends JFrame implements View
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(layout.createSequentialGroup()
-                        .addComponent(guessedLettersLabel)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(guessedLetters, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(layout.createSequentialGroup()
-                                .addComponent(guess, javax.swing.GroupLayout.PREFERRED_SIZE, 128, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                .addComponent(jButton1))
-                            .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 301, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(wordToGuess, javax.swing.GroupLayout.PREFERRED_SIZE, 250, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(drawPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addComponent(gamePanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(30, 30, 30)
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(drawPanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGroup(layout.createSequentialGroup()
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(guessedLettersLabel)
-                            .addComponent(guessedLetters))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(guess, javax.swing.GroupLayout.PREFERRED_SIZE, 33, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jButton1))
-                        .addGap(18, 18, 18)
-                        .addComponent(wordToGuess)
-                        .addGap(17, 17, 17)
-                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 110, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addContainerGap()
+                .addComponent(gamePanel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
         pack();
     }// </editor-fold>//GEN-END:initComponents
-
-    private void guessFocusGained(java.awt.event.FocusEvent evt)//GEN-FIRST:event_guessFocusGained
-    {//GEN-HEADEREND:event_guessFocusGained
-        guess.selectAll();
-    }//GEN-LAST:event_guessFocusGained
 
     /**
      * @param args the command line arguments
@@ -195,86 +107,12 @@ public class HangmanFrame extends JFrame implements View
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
-    private com.mycompany.hangman.drawing.DrawPanel drawPanel;
-    private javax.swing.JTextField guess;
-    private javax.swing.JLabel guessedLetters;
-    private javax.swing.JLabel guessedLettersLabel;
-    private javax.swing.JButton jButton1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JTextArea outputConsole;
-    private javax.swing.JLabel wordToGuess;
+    private com.mycompany.hangman.gui.GamePanel gamePanel;
     // End of variables declaration//GEN-END:variables
 
-
-    public void modelPropertyChange(PropertyChangeEvent evt)
+    public GamePanel getGamePanel()
     {
-            if (evt.getPropertyName().equals(HangmanGame.OUT_TEXT))
-            {
-                for (Object object : ((Iterable)evt.getNewValue()))
-                {
-                    println(object.toString());
-                }
-            }
-            else if (evt.getPropertyName().equals(HangmanGame.CLEAR_OUT_TEXT))
-            {
-                clearOutputConsole();
-            }
-            else if (evt.getPropertyName().equals(HangmanGame.GUESSED_LETTER))
-            {
-                    setGuessedLetters((List<Character>)evt.getNewValue());
-                    getDrawPanel().repaint();
-            }
-            else if (evt.getPropertyName().equals(HangmanGame.WORD))
-            {
-                    setWordToGuess(evt.getNewValue().toString());
-            }
-    }
-
-    public void setGuessedLetters(List<Character> guessedLetters)
-    {
-        StringBuilder builder = new StringBuilder();
-        for (Character guessedLetter : guessedLetters)
-        {
-            builder.append(guessedLetter);
-        }
-        this.guessedLetters.setText(builder.toString());
-
-    }
-
-    public void setWordToGuess(String string)
-    {
-        this.wordToGuess.setText(string);
-    }
-
-    private void println(String stringToPrint)
-    {
-        outputConsole.append(stringToPrint+"\n");
-    }
-
-    private void clearOutputConsole()
-    {
-        outputConsole.setText("");
-    }
-
-    public String getGuess()
-    {
-        return guess.getText();
-    }
-
-    public void clearGuess()
-    {
-        guess.setText("");
-    }
-
-    public void addController(ActionListener listener)
-    {
-        guess.addActionListener(listener);
-        jButton1.addActionListener(listener);
-    }
-
-    public DrawPanel getDrawPanel()
-    {
-        return drawPanel;
+        return gamePanel;
     }
 
 }
